+++ .firebaserc
@@ -1,5 +1,5 @@
 {
   "projects": {
-    "default": "resumator-bd244"
+    "default": "fm-resume"
   }
 }
